Move wrapper code into a subdirectory+fix UT
We're going to be adding more wrapper code. Makes sense to put it into
its own subdirectory so we can easily view it all.

At the same time, fix up a UT that was failing due to some mock changes.
