+++ source/common/http/wrapped/wrapped_connection_pool.h
