+++ source/common/upstream/cluster_manager_impl.cc
@@ -22,10 +22,10 @@
 #include "common/filesystem/filesystem_impl.h"
 #include "common/grpc/async_client_manager_impl.h"
 #include "common/http/async_client_impl.h"
-#include "common/http/connection_mapper_factory.h"
 #include "common/http/http1/conn_pool.h"
 #include "common/http/http2/conn_pool.h"
-#include "common/http/wrapped_connection_pool.h"
+#include "common/http/wrapped/connection_mapper_factory.h"
+#include "common/http/wrapped/wrapped_connection_pool.h"
 #include "common/json/config_schemas.h"
 #include "common/network/resolver_impl.h"
 #include "common/network/utility.h"
